requirements.txt
streamlit
google-api-python-client
python-dotenv
requests
duckduckgo-search